smart-expense-tracker/requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
WTForms==3.1.2
email-validator==2.1.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0
gunicorn==22.0.0
reportlab==4.2.5
